Created a ping page to demonstrate how to get JSON data from pages;
