+++ server/app/ping.php
@@ -1,0 +1,11 @@
+<?php
+require 'FileIO.php';
+
+##$filepath = "/var/www/ProjectJello/Data/test.txt";
+
+##$read_data = read_file($filepath, $squirrel);
+
+echo '{ "name":"microtime", "time":'.microtime(true).'}';
+return;
+
+?>
